pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "RESTful API built with FastAPI framework providing endpoints for data access."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.115.4",
    "geopandas==1.0.1",
    "geopy==2.4.1",
    "google-cloud-bigquery==3.26.0",
    "gunicorn==23.0.0",
    "numpy==2.1.3",
    "pandas==2.2.3",
]
